add the final task

+++ 0x15-file_io/3-cp.c
@@ -9,7 +9,7 @@
 
 int main(int ac, char **av)
 {
-	int as;
+	ssize_t as;
 
 	if (ac != 3)
 	{
